+++ src/app/commands/poap/POAP.ts
@@ -238,8 +238,8 @@ module.exports = class poap extends SlashCommand {
 			return;
 		} catch (e) {
 			LogUtils.logError('failed to process POAP command', e);
-			return ctx.send('Welp, something is definitely broken. I would blame you, but I know better. I\'ll let my devs ' +
-				'know something is wrong.');
+			await ServiceUtils.sendOutErrorMessage(ctx);
+			return;
 		}
 	}
 
@@ -253,7 +253,8 @@ module.exports = class poap extends SlashCommand {
 				return;
 			} else {
 				LogUtils.logError('failed to handle poap command', e);
-				return ServiceUtils.sendOutErrorMessage(ctx);
+				await ServiceUtils.sendOutErrorMessage(ctx);
+				return;
 			}
 		});
 	}